+++ client/actions/actions.js
@@ -1,4 +1,4 @@
-import * as types from './actionTypes.js';
+import * as types from './actionTypes.js.js';
 
 // ------ ACTIONS
