+++ client/constants/actionTypes.js
