+++ client/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import userReducers from './reducers.js';
+import userReducers from './reducers/reducers.js/index.js';
 
 export default combineReducers({
   reduxState: userReducers
